@@ -9,7 +9,7 @@ repository: https://github.com/rikulo/socket.io-dart
 issue_tracker: https://github.com/rikulo/socket.io-dart/issues
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   stream: ^3.0.0
