@@ -8,6 +8,7 @@
 ///    17/02/2017, Created by jumperchen
 ///
 /// Copyright (C) 2017 Potix Corporation. All Rights Reserved.
+
 import 'package:logging/logging.dart';
 import 'package:socket_io/src/engine/connect.dart';
 import 'package:socket_io_common/src/engine/parser/parser.dart';
@@ -41,17 +42,17 @@ class Transports {
 
 abstract class Transport extends EventEmitter {
   static final Logger _logger = Logger('socket_io:transport.Transport');
-  double maxHttpBufferSize;
-  Map httpCompression;
-  Map perMessageDeflate;
-  bool supportsBinary;
-  String sid;
-  String name;
-  bool writable;
-  String readyState;
-  bool discarded;
-  SocketConnect connect;
-  MessageHandler messageHandler;
+  double? maxHttpBufferSize;
+  Map? httpCompression;
+  Map? perMessageDeflate;
+  bool? supportsBinary;
+  String? sid;
+  String? name;
+  bool? writable;
+  String? readyState;
+  late bool discarded;
+  SocketConnect? connect;
+  MessageHandler? messageHandler;
 
   Transport(connect) {
     readyState = 'open';
@@ -72,13 +73,13 @@ abstract class Transport extends EventEmitter {
     this.connect = connect;
   }
 
-  void close([dynamic Function() closeFn]) {
+  void close([dynamic Function()? closeFn]) {
     if ('closed' == readyState || 'closing' == readyState) return;
     readyState = 'closing';
     doClose(closeFn);
   }
 
-  void doClose([dynamic Function() callback]);
+  void doClose([dynamic Function()? callback]);
 
   void onError(msg, [desc]) {
     writable = false;
@@ -89,15 +90,15 @@ abstract class Transport extends EventEmitter {
     }
   }
 
-  void onPacket(Map packet) {
+  void onPacket(Map? packet) {
     emit('packet', packet);
   }
 
   void onData(data) {
     if (messageHandler != null) {
-      messageHandler.handle(this, data);
+      messageHandler?.handle(this, data);
     } else {
-      onPacket(PacketParser.decodePacket(data));
+      onPacket(PacketParser.decodePacket(data, utf8decode: true));
     }
   }
 
