@@ -15,6 +15,7 @@ import 'package:socket_io_common/src/parser/parser.dart';
 import 'package:socket_io/src/namespace.dart';
 import 'package:socket_io/src/server.dart';
 import 'package:socket_io/src/util/event_emitter.dart';
+
 /// Module exports.
 //
 //module.exports = exports = Socket;
@@ -49,18 +50,18 @@ class Socket extends EventEmitter {
   // ignore: undefined_class
   Namespace nsp;
   Client client;
-  Server server;
-  Adapter adapter;
-  String id;
-  HttpRequest request;
+  late Server server;
+  Adapter? adapter;
+  late String id;
+  late HttpRequest request;
   var conn;
   Map roomMap = {};
   List roomList = [];
   Map acks = {};
   bool connected = true;
   bool disconnected = false;
-  Map handshake;
-  Map<String, bool> flags;
+  Map? handshake;
+  Map<String, bool> flags = {};
 
   // a data store for each socket.
   Map data = {};
@@ -98,19 +99,16 @@ class Socket extends EventEmitter {
   }
 
   Socket get json {
-    flags = flags ?? {};
     flags['json'] = true;
     return this;
   }
 
   Socket get volatile {
-    flags = flags ?? {};
     flags['volatile'] = true;
     return this;
   }
 
   Socket get broadcast {
-    flags = flags ?? {};
     flags['broadcast'] = true;
     return this;
   }
@@ -129,14 +127,14 @@ class Socket extends EventEmitter {
   /// @return {Socket} self
   /// @api public
   void emitWithAck(String event, dynamic data,
-      {Function ack, bool binary = false}) {
+      {Function? ack, bool binary = false}) {
     if (EVENTS.contains(event)) {
       super.emit(event, data);
     } else {
       var packet = {};
       var sendData = data == null ? [event] : [event, data];
 
-      var flags = this.flags ?? {};
+      var flags = this.flags;
 
       if (ack != null) {
         if (roomList.isNotEmpty || flags['broadcast'] == true) {
@@ -152,7 +150,7 @@ class Socket extends EventEmitter {
       packet['data'] = sendData;
 
       if (roomList.isNotEmpty || flags['broadcast'] == true) {
-        adapter.broadcast(packet, {
+        adapter!.broadcast(packet, {
           'except': [id],
           'rooms': roomList,
           'flags': flags
@@ -165,7 +163,7 @@ class Socket extends EventEmitter {
 
 //      // reset flags
       roomList = [];
-      this.flags = null;
+      this.flags = {};
 //    }
 //    return this;
     }
@@ -221,7 +219,7 @@ class Socket extends EventEmitter {
       if (fn != null) fn(null);
       return this;
     }
-    adapter.add(id, room, ([err]) {
+    adapter!.add(id, room, ([err]) {
       if (err != null) return fn?.call(err);
 //      _logger.info('joined room %s', room);
       roomMap[room] = room;
@@ -238,7 +236,7 @@ class Socket extends EventEmitter {
   /// @api private
   Socket leave(room, fn) {
 //    debug('leave room %s', room);
-    adapter.del(id, room, ([err]) {
+    adapter!.del(id, room, ([err]) {
       if (err != null) return fn?.call(err);
 //      _logger.info('left room %s', room);
       roomMap.remove(room);
@@ -252,7 +250,7 @@ class Socket extends EventEmitter {
   /// @api private
 
   void leaveAll() {
-    adapter.delAll(id);
+    adapter!.delAll(id);
     roomMap = {};
   }
 
@@ -348,7 +346,7 @@ class Socket extends EventEmitter {
   ///
   /// @api private
   void onack(packet) {
-    Function ack = acks.remove(packet['id']);
+    Function? ack = acks.remove(packet['id']);
     if (ack is Function) {
 //      debug('calling ack %s with %j', packet.id, packet.data);
       Function.apply(ack, packet['data']);
@@ -426,7 +424,6 @@ class Socket extends EventEmitter {
   /// @return {Socket} self
   /// @api public
   Socket compress(compress) {
-    flags = flags ?? {};
     flags['compress'] = compress;
     return this;
   }
